requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
boto3==1.28.30
botocore==1.31.30
certifi==2023.7.22
deprecation==2.1.0
Django==4.2.4
django-indexer==0.3.0
django-paging==0.2.5
django-phonenumber-field==7.1.0
django-redis==5.3.0
django-s3direct==2.0.3
django-storages==1.13.2
django-templatetag-sugar==1.0
djangorestframework==3.14.0
dnspython==2.4.2
docutils==0.20.1
eventlet==0.33.3
exceptiongroup==1.1.3
gotrue==1.0.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
jmespath==1.0.1
lockfile==0.12.2
packaging==23.1
postgrest==0.10.8
psycopg2==2.9.7
pydantic==2.2.1
pydantic_core==2.6.1
python-daemon==3.0.1
python-dateutil==2.8.2
pytz==2023.3
raven==6.10.0
realtime==1.0.0
redis==5.0.0
s3transfer==0.6.2
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
South==1.0.2
sqlparse==0.4.4
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
typing_extensions==4.7.1
urllib3==1.26.16
websockets==10.4
